+++ src/components/spartan/SStackedList/types.ts
@@ -1,0 +1,9 @@
+import type { HTMLAttributes } from "vue";
+
+export type TStackedListProps = {
+    class?: HTMLAttributes['class'];
+};
+
+export type TStackedListItemProps = {
+    class?: HTMLAttributes['class'];
+};