+++ src/components/spartan/index.ts
@@ -48,3 +48,4 @@ export * from './SPageTitle';
 export * from './STextArea';
 export * from './STextAreaBlock';
 export * from './SInputOTP';
+export * from './SStackedList';